Fixed importing to work if importer / hooks are in sep. directories
We used to just pass in the name of the module (i.e. ./some_hooks.py we
would pass in some_hooks), but this only worked if the importer and the
hooks-module were in the same directory. Now you instead pass in the
full path to the python file containing the hooks allowing us to work
across different directories.
Also added '.processes.shelve' to the .gitignore, this is a temp file
used by the importer.